+++ src/app/controller/peopleController.js
@@ -8,7 +8,10 @@ class PeopleController{
             const newPeople = await peopleService.create(req.body);
             return res.status(201).json({"Pessoa":newPeople});
         } catch (error) {
-            return res.status(500).json({error});
+            return res.status(400).json({
+            'message': 'bad request',
+            'details': [{ 'message': error.message }]
+            });
         }
     }
 
@@ -18,7 +21,10 @@ class PeopleController{
             return res.status(200).json(allPeoples);
 
         } catch (error) {
-            return res.status(500).json({error});
+            return res.status(400).json({
+            'message': 'bad request',
+            'details': [{ 'message': error.message }]
+            });
             
         }
     }
@@ -34,7 +40,10 @@ class PeopleController{
             
             return res.status(200).json({"Pessoa":onePeople});
         } catch (error) {
-           return res.status(500).json({error});
+           return res.status(500).json({
+            'message': 'bad request',
+            'details': [{ 'message': error.message }]
+           });
             
         }
     }
@@ -52,7 +61,10 @@ class PeopleController{
             const  updatedPeople = await peopleService.updateOne(id,reqPeople);
             return res.status(200).json(updatedPeople);
         } catch (error) {
-            return res.status(400).json({error});
+            return res.status(400).json({ 
+            'message': 'bad request',
+            'details': [{ 'message': error.message }]
+            });
             
         }
     }
@@ -71,7 +83,10 @@ class PeopleController{
             return res.status(204).json();
 
         } catch (error) {
-           return res.status(500).json({error});   
+           return res.status(400).json({  
+            'message': 'bad request',
+            'details': [{ 'message': error.message }]
+           });   
         }
     }
